refactor toolbar dialogs
