+++ packages/lexical-playground/src/plugins/AutoEmbedPlugin/index.tsx
@@ -22,6 +22,7 @@ import * as ReactDOM from 'react-dom';
 
 import useModal from '../../hooks/useModal';
 import Button from '../../ui/Button';
+import {DialogActions} from '../../ui/Dialog';
 import {INSERT_FIGMA_COMMAND} from '../FigmaPlugin';
 import {INSERT_TWEET_COMMAND} from '../TwitterPlugin';
 import {INSERT_YOUTUBE_COMMAND} from '../YouTubePlugin';
@@ -248,14 +249,14 @@ export function AutoEmbedDialog({
           }}
         />
       </div>
-      <div className="ToolbarPlugin__dialogActions">
+      <DialogActions>
         <Button
           disabled={!embedResult}
           onClick={onClick}
           data-test-id={`${embedConfig.type}-embed-modal-submit-btn`}>
           Embed
         </Button>
-      </div>
+      </DialogActions>
     </div>
   );
 }
