+++ packages/lexical-playground/src/plugins/PollPlugin/index.tsx
@@ -15,12 +15,42 @@ import {
   COMMAND_PRIORITY_EDITOR,
   createCommand,
   LexicalCommand,
+  LexicalEditor,
 } from 'lexical';
-import {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 
 import {$createPollNode, PollNode} from '../../nodes/PollNode';
+import Button from '../../ui/Button';
+import {DialogActions} from '../../ui/Dialog';
+import TextInput from '../../ui/TextInput';
 
 export const INSERT_POLL_COMMAND: LexicalCommand<string> = createCommand();
+
+export function InsertPollDialog({
+  activeEditor,
+  onClose,
+}: {
+  activeEditor: LexicalEditor;
+  onClose: () => void;
+}): JSX.Element {
+  const [question, setQuestion] = useState('');
+
+  const onClick = () => {
+    activeEditor.dispatchCommand(INSERT_POLL_COMMAND, question);
+    onClose();
+  };
+
+  return (
+    <>
+      <TextInput label="Question" onChange={setQuestion} value={question} />
+      <DialogActions>
+        <Button disabled={question.trim() === ''} onClick={onClick}>
+          Confirm
+        </Button>
+      </DialogActions>
+    </>
+  );
+}
 
 export default function PollPlugin(): JSX.Element | null {
   const [editor] = useLexicalComposerContext();
