--- packages/lexical-playground/src/plugins/ToolbarPlugin/index.css
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- *
- */
-.ToolbarPlugin__dialogActions {
-  display: flex;
-  flex-direction: row;
-  justify-content: right;
-  margin-top: 20px;
-}
-
-.ToolbarPlugin__dialogButtonsList {
-  display: flex;
-  flex-direction: column;
-  justify-content: right;
-  margin-top: 20px;
-}
-
-.ToolbarPlugin__dialogButtonsList button {
-  margin-bottom: 20px;
-}
